temporarly disable parties compatibility
This is required to get an initial JitPack artifact of this plugin as there is no Parties artifact that JitPack can use yet.
Then this initial artifact must be added to the Parties pom to get an initial artifact of Parties on JitPack.
This allows us to add parties as a JitPack dependency for this project and revert this commit.